refactor(LottoResult): 숫자 타입 더 큰 범위 가능하도록 BigInteger, BigDecimal로 수정

+++ src/main/java/lotto/domain/LottoResult.java
@@ -1,12 +1,14 @@
 package lotto.domain;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Map;
 
 public class LottoResult {
     private final Map<Lotto, LottoRank> ranks;
     private final Long totalCost;
-    private final Long totalPrize;
-    private final Double earningsRatio;
+    private final BigInteger totalPrize;
+    private final BigDecimal earningsRatio;
 
     public LottoResult(Map<Lotto, LottoRank> ranks) {
         this.ranks = ranks;
@@ -23,11 +25,11 @@ public class LottoResult {
         return ranks;
     }
 
-    public Long getTotalPrize() {
+    public BigInteger getTotalPrize() {
         return totalPrize;
     }
 
-    public Double getEarningsRatio() {
+    public BigDecimal getEarningsRatio() {
         return earningsRatio;
     }
 }
